Enhancement Hello, I made some fixes to Thus GTP/UEFI support
- Creation of partitions in a gpt disk
- Format EFI partition, and mount it on /boot/efi
- grub2-efi installation

grub2-efi installation isn't done in a chroot because when efibootmgr
runs it doesn't detect a uefi environment and fails to add a new uefi
boot entry.

I tested all installation modes, and the only thing that doesn't
work is encryption without lvm.

Also another problem I found is if the disk is formatted in mbr, when
you create the gpt partition you need to reboot for the kernel to
properly recognize the new disk and be able to create partitions
without problems.

Ramon Buldó